requirements.txt
python-telegram-bot[ext]>=20.0
phonenumbers>=8.13.0
httpx>=0.25.0
python-dotenv>=1.0.0
